fix(package): fix scoped and nested deps
